// src/components/navigation/my-slider/types.ts
/*
 *  Copyright (c) 2025 TechDivision GmbH
 *  All rights reserved
 *
 *  This product includes proprietary software developed at TechDivision GmbH, Germany
 *  For more information see https://www.techdivision.com/
 *
 *  To obtain a valid license for using this software please contact us at
 *  license@techdivision.com
 *
 *  @author Dominik Böck <d.boeck@techdivision.com>
 */
export type SlideModel = {
    itemImageSrc: string;
    thumbnailImageSrc: string;
    alt: string;
    title: string;
};